refactor(service): move arquivos de exercise para o novo diretorio evaluation
